Finalizando os recursos de pacientes
