+++ src/main/java/br/com/mindsy/api/gateway/controller/AuthenticationController.java
@@ -1,10 +1,9 @@
 package br.com.mindsy.api.gateway.controller;
 
-import br.com.mindsy.api.gateway.dto.AuthenticationRequestDto;
-import br.com.mindsy.api.gateway.dto.AuthenticationResponseDto;
+import br.com.mindsy.api.gateway.dto.authentication.AuthenticationRequestDto;
+import br.com.mindsy.api.gateway.dto.authentication.AuthenticationResponseDto;
 import br.com.mindsy.api.gateway.dto.MessageResponseDto;
 import br.com.mindsy.api.gateway.exception.ApiGatewayException;
-import br.com.mindsy.api.gateway.exception.ObjectAlredyExistsException;
 import br.com.mindsy.api.gateway.exception.UnauthorizadExeption;
 import br.com.mindsy.api.gateway.service.AuthenticationService;
 import br.com.mindsy.api.gateway.service.PsychologistService;
@@ -28,7 +27,7 @@ public class AuthenticationController {
     }
 
     @PutMapping("/logout/{crp}")
-    public MessageResponseDto logout(@PathVariable("crp") final String crp, @RequestHeader("Authorization") final String bearerToken) throws UnauthorizadExeption, ApiGatewayException, ObjectAlredyExistsException {
+    public MessageResponseDto logout(@PathVariable("crp") final String crp, @RequestHeader("Authorization") final String bearerToken) throws UnauthorizadExeption, ApiGatewayException {
         psychologistService.validateToken(crp, bearerToken);
         return authenticationService.logout(bearerToken, crp);
     }
