+++ src/main/java/br/com/mindsy/api/gateway/controller/PatientController.java
@@ -1,8 +1,8 @@
 package br.com.mindsy.api.gateway.controller;
 
 import br.com.mindsy.api.gateway.dto.MessageResponseDto;
-import br.com.mindsy.api.gateway.dto.PatientRequestDto;
-import br.com.mindsy.api.gateway.dto.PatientResponseDto;
+import br.com.mindsy.api.gateway.dto.patient.PatientRequestDto;
+import br.com.mindsy.api.gateway.dto.patient.PatientResponseDto;
 import br.com.mindsy.api.gateway.exception.ApiGatewayException;
 import br.com.mindsy.api.gateway.exception.ObjectAlredyExistsException;
 import br.com.mindsy.api.gateway.exception.UnauthorizadExeption;
@@ -12,6 +12,7 @@ import br.com.mindsy.api.gateway.service.PsychologistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -25,7 +26,7 @@ public class PatientController {
     PsychologistService psychologistService;
 
     @PostMapping
-    public MessageResponseDto insert(@RequestBody PatientRequestDto patientRequestDto,
+    public MessageResponseDto insert(@Valid @RequestBody PatientRequestDto patientRequestDto,
                                      @RequestHeader("Authorization") final String bearerToken,
                                      @RequestHeader("crp") final String crp) throws ApiGatewayException, ObjectAlredyExistsException, UnauthorizadExeption {
         psychologistService.validateToken(crp, bearerToken);
@@ -35,15 +36,24 @@ public class PatientController {
     @GetMapping("{id}")
     public PatientResponseDto get(@PathVariable("id") final Long id,
                                      @RequestHeader("Authorization") final String bearerToken,
-                                     @RequestHeader("crp") final String crp) throws ApiGatewayException, ObjectAlredyExistsException, UnauthorizadExeption, UserNotFoundException {
+                                     @RequestHeader("crp") final String crp) throws ApiGatewayException, UnauthorizadExeption, UserNotFoundException {
         psychologistService.validateToken(crp, bearerToken);
         return patientService.get(id);
     }
 
     @GetMapping("/psychologist/{crp}")
     public List<PatientResponseDto> get(@PathVariable("crp") final String crp,
-                                        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, ObjectAlredyExistsException, UnauthorizadExeption, UserNotFoundException {
+                                        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UnauthorizadExeption, UserNotFoundException {
         psychologistService.validateToken(crp, bearerToken);
         return patientService.getByCrp(crp);
+    }
+
+    @PutMapping("/{id}")
+    public MessageResponseDto update(@PathVariable("id") final Long id,
+                                           @RequestBody PatientRequestDto patientRequestDto,
+                                           @RequestHeader("crp") final String crp,
+                                           @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UnauthorizadExeption, UserNotFoundException {
+        psychologistService.validateToken(crp, bearerToken);
+        return patientService.update(id, patientRequestDto);
     }
 }
