+++ src/main/java/br/com/mindsy/api/gateway/controller/PsychologistController.java
@@ -1,8 +1,8 @@
 package br.com.mindsy.api.gateway.controller;
 
-import br.com.mindsy.api.gateway.dto.PsychologistRequestDto;
+import br.com.mindsy.api.gateway.dto.psychologist.PsychologistRequestDto;
 import br.com.mindsy.api.gateway.dto.MessageResponseDto;
-import br.com.mindsy.api.gateway.dto.PsychologistResponseDto;
+import br.com.mindsy.api.gateway.dto.psychologist.PsychologistResponseDto;
 import br.com.mindsy.api.gateway.exception.*;
 import br.com.mindsy.api.gateway.service.PsychologistService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,21 +22,21 @@ public class PsychologistController {
 
     @PutMapping("/{crp}")
     public MessageResponseDto update(@PathVariable("crp") final String crp, @RequestBody PsychologistRequestDto psychologistRequestDto,
-                                     @RequestHeader("Authorization") final String bearerToken) throws InvalidParameterException, ApiGatewayException, UserNotFoundException, ObjectAlredyExistsException, UnauthorizadExeption {
+                                     @RequestHeader("Authorization") final String bearerToken) throws InvalidParameterException, ApiGatewayException, UserNotFoundException, UnauthorizadExeption {
         psychologistService.validateToken(crp, bearerToken);
         return psychologistService.update(crp, psychologistRequestDto);
     }
 
     @DeleteMapping("/{crp}")
     public MessageResponseDto delete(@PathVariable("crp") final String crp,
-        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UserNotFoundException, ObjectAlredyExistsException, UnauthorizadExeption {
+        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UserNotFoundException, UnauthorizadExeption {
         psychologistService.validateToken(crp, bearerToken);
         return psychologistService.delete(crp);
     }
 
     @GetMapping("/{crp}")
     public PsychologistResponseDto find(@PathVariable("crp") final String crp,
-                                        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UserNotFoundException, ObjectAlredyExistsException, UnauthorizadExeption {
+                                        @RequestHeader("Authorization") final String bearerToken) throws ApiGatewayException, UserNotFoundException, UnauthorizadExeption {
         psychologistService.validateToken(crp, bearerToken);
         return psychologistService.find(crp);
     }
