+++ src/main/java/br/com/mindsy/api/gateway/dto/authentication/AuthenticationRequestDto.java
@@ -1,13 +1,12 @@
-package br.com.mindsy.api.gateway.dto;
+package br.com.mindsy.api.gateway.dto.authentication;
 
 import java.io.Serializable;
 
-public class PsychologistDto implements Serializable {
+public class AuthenticationRequestDto implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
     private String crp;
-    private String date_of_birth;
     private String password;
 
     public String getCrp() {
@@ -16,14 +15,6 @@ public class PsychologistDto implements Serializable {
 
     public void setCrp(String crp) {
         this.crp = crp;
-    }
-
-    public String getDate_of_birth() {
-        return date_of_birth;
-    }
-
-    public void setDate_of_birth(String date_of_birth) {
-        this.date_of_birth = date_of_birth;
     }
 
     public String getPassword() {
