+++ src/main/java/br/com/mindsy/api/gateway/dto/psychologist/PsychologistBackEndDto.java
@@ -1,4 +1,4 @@
-package br.com.mindsy.api.gateway.dto;
+package br.com.mindsy.api.gateway.dto.psychologist;
 
 import java.io.Serializable;
