+++ src/main/java/br/com/mindsy/api/gateway/dto/psychologist/PsychologistDto.java
@@ -1,12 +1,13 @@
-package br.com.mindsy.api.gateway.dto;
+package br.com.mindsy.api.gateway.dto.psychologist;
 
 import java.io.Serializable;
 
-public class AuthenticationRequestDto implements Serializable {
+public class PsychologistDto implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
     private String crp;
+    private String date_of_birth;
     private String password;
 
     public String getCrp() {
@@ -15,6 +16,14 @@ public class AuthenticationRequestDto implements Serializable {
 
     public void setCrp(String crp) {
         this.crp = crp;
+    }
+
+    public String getDate_of_birth() {
+        return date_of_birth;
+    }
+
+    public void setDate_of_birth(String date_of_birth) {
+        this.date_of_birth = date_of_birth;
     }
 
     public String getPassword() {
