+++ src/main/java/br/com/mindsy/api/gateway/dto/psychologist/PsychologistRequestDto.java
@@ -1,4 +1,4 @@
-package br.com.mindsy.api.gateway.dto;
+package br.com.mindsy.api.gateway.dto.psychologist;
 
 import br.com.mindsy.api.gateway.dto.enums.PhoneTypeEnum;
