+++ src/main/java/br/com/mindsy/api/gateway/service/feign/AuthenticationFeign.java
@@ -1,7 +1,7 @@
 package br.com.mindsy.api.gateway.service.feign;
 
-import br.com.mindsy.api.gateway.dto.AuthenticationRequestDto;
-import br.com.mindsy.api.gateway.dto.AuthenticationResponseDto;
+import br.com.mindsy.api.gateway.dto.authentication.AuthenticationRequestDto;
+import br.com.mindsy.api.gateway.dto.authentication.AuthenticationResponseDto;
 import br.com.mindsy.api.gateway.dto.MessageResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.stereotype.Service;
