+++ src/main/java/br/com/mindsy/api/gateway/service/feign/PatientFeign.java
@@ -1,14 +1,11 @@
 package br.com.mindsy.api.gateway.service.feign;
 
 import br.com.mindsy.api.gateway.dto.MessageResponseDto;
-import br.com.mindsy.api.gateway.dto.PatientBackendDto;
-import br.com.mindsy.api.gateway.dto.PatientResponseBackDto;
+import br.com.mindsy.api.gateway.dto.patient.PatientBackendDto;
+import br.com.mindsy.api.gateway.dto.patient.PatientResponseBackDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -24,4 +21,7 @@ public interface PatientFeign {
 
     @GetMapping("/patient/psychologist/{crp}")
     List<PatientResponseBackDto> getByCrp(@PathVariable("crp") final String id);
+
+    @PutMapping("/patient/{id}")
+    MessageResponseDto update(@PathVariable("id") final Long id, final PatientBackendDto patientBackendDto);
 }
